requirements.txt
beautifulsoup4==4.13.5
httpx==0.28.1
pipreqs==0.4.13
pyside6==6.9.2
pyside6_addons==6.9.2
pyside6_essentials==6.9.2
python-dotenv==1.1.1
Requests==2.32.5
selenium==4.35.0
selenium_stealth==1.0.6
tqdm==4.66.6
webdriver_manager==4.0.2
rapidfuzz==3.9.4
rich==13.7.1

# Development dependencies
pytest
ruff
mypy